+++ src/components/Posts/style.ts
@@ -60,3 +60,8 @@ export const Hr = styled.hr`
   border-top: none;
   border-bottom: 0.5px solid lightgray;
 `;
+
+export const ErrDiv = styled.div`
+  display: flex;
+  justify-content: center;
+`;
