+++ src/pages/Search/type.tsx
@@ -1,0 +1,7 @@
+export interface ProdData {
+  product_code: number;
+  name: string;
+  image_url: string;
+  price: number;
+  category_names: string[];
+}
